Excel file containing list of student details
This file will contain the list of students with their details such as Name, Reg No and kmail ID. This file is then uploaded into the Admin Portal by the Admin, and the backend server will then send a registration link to the enlisted students via their kmail IDs, which they can use to register an account in the Student Portal. Do not change the attribute names of this excel sheet, as the server looks out for those specific attribute names.